Updated the adjoint electroatom factory and adjoint electroatomic reaction factory classes to set the critical line energies.
When the adjoint bremsstrahlung and electro-ionization classes are now created with the critical line energies set by the user in the adjoint electron simulation properties. aprobinson/FRENSIE#164

+++ packages/monte_carlo/collision/electron/src/MonteCarlo_ElectroatomNativeFactory_def.hpp
@@ -181,7 +181,7 @@ void ElectroatomNativeFactory::createElectroatomCore(
                       reaction_pointers,
                       properties.getElectronEvaluationTolerance() );
 
-    for( unsigned i = 0; i < reaction_pointers.size(); ++i )
+    for( size_t i = 0; i < reaction_pointers.size(); ++i )
     {
       scattering_reactions[reaction_pointers[i]->getReactionType()] =
         reaction_pointers[i];
